+++ lib/src/constant/month.dart
@@ -1,5 +1,11 @@
-// 月：2020-10
+// 年月：2020-10
 const MONTH_DEFAULT = 'yyyy-MM';
 
-// 中文版
+// 只有月：10
+const MONTH_ONLY = 'MM';
+
+// 中文版年月
 const MONTH_CHINESE = 'yyyy年M月';
+
+// 中文版只有月
+const MONTH_ONLY_CHINESE = 'M月';