+++ CHANGELOG.md
@@ -92,3 +92,4 @@
 - 新增 hexToRgbaObject, hexToRgbaString
 - 新增 getStringLength, trimString, sliceString, randomString, renderStringTemplate
 - 新增 maskMobile, maskName, randomIntegerByLength, randomIntegerByRange, randomStringByLength, encodeURIComponent, decodeURIComponent
+- 新增 maskEmail, padStringStart